V16 without synchronized pools
