+++ src/main/java/ch/ethz/globis/phtree/v13/Bits.java
@@ -1,14 +1,25 @@
 /*
  * Copyright 2011-2016 ETH Zurich. All Rights Reserved.
  * Copyright 2016-2018 Tilmann Zäschke. All Rights Reserved.
+ * Copyright 2019 Improbable. All rights reserved.
  *
- * This software is the proprietary information of ETH Zurich
- * and Tilmann Zäschke.
- * Use is subject to license terms.
+ * This file is part of the PH-Tree project.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package ch.ethz.globis.phtree.v13;
 
-import ch.ethz.globis.phtree.util.BitsLong;
+import ch.ethz.globis.phtree.util.unsynced.LongArrayOps;
 
 
 /**
@@ -17,7 +28,7 @@ import ch.ethz.globis.phtree.util.BitsLong;
  * @author ztilmann
  *
  */
-public class Bits extends BitsLong {
+public class Bits extends LongArrayOps {
 
 
 }
