+++ src/main/java/ch/ethz/globis/phtree/v13/Node.java
@@ -21,17 +21,7 @@ package ch.ethz.globis.phtree.v13;
 
 import static ch.ethz.globis.phtree.PhTreeHelper.posInArray;
 
-import ch.ethz.globis.pht64kd.MaxKTreeI.NtEntry;
-import ch.ethz.globis.pht64kd.MaxKTreeI.PhIterator64;
-import ch.ethz.globis.phtree.PhEntry;
 import ch.ethz.globis.phtree.PhTreeHelper;
-import ch.ethz.globis.phtree.util.Refs;
-import ch.ethz.globis.phtree.util.RefsLong;
-import ch.ethz.globis.phtree.v13.nt.NodeTreeV13;
-import ch.ethz.globis.phtree.v13.nt.NtIteratorMask;
-import ch.ethz.globis.phtree.v13.nt.NtIteratorMinMax;
-import ch.ethz.globis.phtree.v13.nt.NtNode;
-import ch.ethz.globis.phtree.v13.nt.NtNodePool;
 
 
 /**
@@ -47,7 +37,6 @@ public class Node {
 	private static final int INN_HC_WIDTH = 0; //Index-NotNull: width of not-null flag for post/infix-hc
 	/** Bias towards using AHC. AHC is used if (sizeLHC*AHC_LHC_BIAS) greater than (sizeAHC)  */
 	public static final double AHC_LHC_BIAS = 2.0; 
-	public static final int NT_THRESHOLD = 150; 
 
 	private Object[] values;
 	
@@ -82,43 +71,23 @@ public class Node {
 	private byte postLenStored = 0;
 	private byte infixLenStored = 0; //prefix size
 
-	//Nested tree index
-	private NtNode<Object> ind = null;
-
 	
-	/**
-	 * @return true if NI should be used. 
-	 */
-	private static final boolean shouldSwitchToNT(int entryCount) {
-		//Maybe just provide a switching threshold? 5-10?
-		return entryCount >= NT_THRESHOLD;
-	}
-
-	private static final boolean shouldSwitchFromNtToHC(int entryCount) {
-		return entryCount <= NT_THRESHOLD-30;
-	}
-
-	static final int IK_WIDTH(int dims) { return dims; }; //post index key width 
+	static final int IK_WIDTH(int dims) { return dims; } //post index key width
 
     private Node() {
 		// For ZooDB only
 	}
 
-	protected Node(Node original) {
+	protected Node(Node original, PhTree13<?> tree) {
         if (original.values != null) {
-            this.values = Refs.arrayClone(original.values);
+            this.values = tree.objPool().arrayClone(original.values);
         }
         this.entryCnt = original.entryCnt;
-        this.infixLenStored = original.infixLenStored;
         this.isHC = original.isHC;
         this.postLenStored = original.postLenStored;
         this.infixLenStored = original.infixLenStored;
-        if (original.ind != null) {
-        	//copy NT tree
-        	throw new UnsupportedOperationException();
-        }
         if (original.ba != null) {
-        	this.ba = Bits.arrayClone(original.ba);
+        	this.ba = tree.longPool().arrayClone(original.ba);
         }
     }
 
@@ -126,37 +95,27 @@ public class Node {
 		return new Node();
 	}
 
-	private void initNode(int infixLenClassic, int postLenClassic, int dims) {
+	private void initNode(int infixLenClassic, int postLenClassic, int dims, PhTree13<?> tree) {
 		this.infixLenStored = (byte) (infixLenClassic + 1);
 		this.postLenStored = (byte) (postLenClassic + 1);
 		this.entryCnt = 0;
-		this.ind = null;
 		this.isHC = false;
 		int size = calcArraySizeTotalBits(2, dims);
-		this.ba = Bits.arrayCreate(size);
-		this.values = Refs.arrayCreate(2);
+		this.ba = tree.longPool().arrayCreate(size);
+		this.values = tree.objPool().arrayCreate(2);
 	}
 
-	static Node createNode(int dims, int infixLenClassic, int postLenClassic) {
-		Node n = NodePool.getNode();
-		n.initNode(infixLenClassic, postLenClassic, dims);
+	static Node createNode(int dims, int infixLenClassic, int postLenClassic, PhTree13<?> tree) {
+		Node n = tree.nodePool().get();
+		n.initNode(infixLenClassic, postLenClassic, dims, tree);
 		return n;
 	}
 
-	static Node createNode(Node original) {
-		return new Node(original);
-	}
-
-	<T> PhEntry<T> createNodeEntry(long[] key, T value) {
-		return new PhEntry<>(key, value);
-	}
-	
-	void discardNode() {
-		Bits.arrayReplace(ba, null);
-		Refs.arrayReplace(values, null);
+	private void discardNode(PhTree13<?> tree) {
+		tree.longPool().arrayReplace(ba, null);
+		tree.objPool().arrayReplace(values, null);
 		entryCnt = 0;
-		NodePool.offer(this);
-		ind = null;
+		tree.nodePool().offer(this);
 	}
 	
 	int calcArraySizeTotalBits(int entryCount, final int dims) {
@@ -165,19 +124,12 @@ public class Node {
 		if (isAHC()) {
 			//hyper-cube
 			nBits += (INN_HC_WIDTH + dims * postLenStored()) * (1 << dims);
-		} else if (isNT()) {
-			nBits += 0;
 		} else {
 			//hc-pos index
 			nBits += entryCount * (IK_WIDTH(dims) + dims * postLenStored());
 		}
 		return nBits;
 	}
-
-	private int calcArraySizeTotalBitsNt() {
-		return getBitPosIndex();
-	}
-
 
 	/**
 	 * 
@@ -186,7 +138,7 @@ public class Node {
 	 * @param outVal output
 	 * @return whether the infix length is > 0
 	 */
-	boolean getInfixOfSub(int pin, long hcPos, long[] outVal) {
+	private boolean getInfixOfSub(int pin, long hcPos, long[] outVal) {
 		int offs = pinToOffsBitsData(pin, hcPos, outVal.length);
 		if (!hasSubInfix(offs, outVal.length)) {
 			applyHcPos(hcPos, outVal);
@@ -211,22 +163,11 @@ public class Node {
 	 */
 	Object doInsertIfMatching(long[] keyToMatch, Object newValueToInsert, PhTree13<?> tree) {
 		long hcPos = posInArray(keyToMatch, getPostLen());
-
-		if (isNT()) {
-			//ntPut will also increase the node-entry count
-			Object v = ntPut(hcPos, keyToMatch, newValueToInsert);
-			//null means: Did not exist, or we had to do a split...
-			if (v == null) {
-				tree.increaseNrEntries();
-			}
-			return v;
-		}
-		
 		int pin = getPosition(hcPos, keyToMatch.length);
 		//check whether hcPos is valid
 		if (pin < 0) {
 			tree.increaseNrEntries();
-			addPostPIN(hcPos, pin, keyToMatch, newValueToInsert);
+			addPostPIN(hcPos, pin, keyToMatch, newValueToInsert, tree);
 			return null;
 		}
 		
@@ -275,22 +216,7 @@ public class Node {
 		
 		long hcPos = posInArray(keyToMatch, getPostLen());
 		
-		if (isNT()) {
-			if (getOnly) {
-				return ntGetEntryIfMatches(hcPos, keyToMatch);
-			}			
-			Object v = ntRemoveEntry(hcPos, keyToMatch, newKey, insertRequired);
-			if (v != null && !(v instanceof Node)) {
-				//Found and removed entry.
-				tree.decreaseNrEntries();
-				if (getEntryCount() == 1) {
-					mergeIntoParentNt(keyToMatch, parent);
-				}
-			}
-			return v;
-		}
-		
-		int pin; 
+		int pin;
 		Object v;
 		int offs;
 		int dims = keyToMatch.length;
@@ -345,11 +271,11 @@ public class Node {
 		return true;
 	}
 
-	public long calcPostfixMask() {
+	private long calcPostfixMask() {
 		return ~((-1L)<<getPostLen());
 	}
 	
-	public long calcInfixMask(int subPostLen) {
+	private long calcInfixMask(int subPostLen) {
 		long mask = ~((-1L)<<(getPostLen()-subPostLen-1));
 		return mask << (subPostLen+1);
 	}
@@ -369,7 +295,7 @@ public class Node {
 	 * @return The value
 	 */
 	private Object insertSplit(long[] newKey, Object newValue, Object currentValue,
-			int pin, long hcPos, PhTree13<?> tree, int offs, long mask) {
+							   int pin, long hcPos, PhTree13<?> tree, int offs, long mask) {
         //do the splitting
 
         //What does 'splitting' mean (we assume there is currently a sub-node, in case of a postfix
@@ -395,9 +321,10 @@ public class Node {
 			return currentValue;
 		}
 		
-		Node newNode = createNode(newKey, newValue, buffer, currentValue, maxConflictingBits);
+		Node newNode =
+                createNode(newKey, newValue, buffer, currentValue, maxConflictingBits, tree);
 
-        replaceEntryWithSub(pin, hcPos, newKey, newNode);
+        replaceEntryWithSub(pin, hcPos, newKey, newNode, tree);
         tree.increaseNrEntries();
 		//entry did not exist
         return null;
@@ -412,21 +339,21 @@ public class Node {
      * @param mcb most conflicting bit
      * @return A new node or 'null' if there are no conflicting bits
      */
-    public Node createNode(long[] key1, Object val1, long[] key2, Object val2,
-    		int mcb) {
+    public Node createNode(long[] key1, Object val1, long[] key2, Object val2, int mcb,
+                           PhTree13<?> tree) {
         //determine length of infix
         int newLocalInfLen = getPostLen() - mcb;
         int newPostLen = mcb-1;
-        Node newNode = createNode(key1.length, newLocalInfLen, newPostLen);
+        Node newNode = createNode(key1.length, newLocalInfLen, newPostLen, tree);
 
         long posSub1 = posInArray(key1, newPostLen);
         long posSub2 = posInArray(key2, newPostLen);
         if (posSub1 < posSub2) {
-        	newNode.writeEntry(0, posSub1, key1, val1);
-        	newNode.writeEntry(1, posSub2, key2, val2);
+        	newNode.writeEntry(0, posSub1, key1, val1, tree);
+        	newNode.writeEntry(1, posSub2, key2, val2, tree);
         } else {
-        	newNode.writeEntry(0, posSub2, key2, val2);
-        	newNode.writeEntry(1, posSub1, key1, val1);
+        	newNode.writeEntry(0, posSub2, key2, val2, tree);
+        	newNode.writeEntry(1, posSub1, key1, val1, tree);
         }
         newNode.incEntryCount();
         newNode.incEntryCount();
@@ -440,7 +367,7 @@ public class Node {
      * @return the position of the most significant conflicting bit (starting with 1) or
      * 0 in case of no conflicts.
      */
-    public static int calcConflictingBits(long[] v1, long[] v2, long mask) {
+    private static int calcConflictingBits(long[] v1, long[] v2, long mask) {
 		//long mask = (1l<<node.getPostLen()) - 1l; // e.g. (0-->0), (1-->1), (8-->127=0x01111111)
      	//write all differences to diff, we just check diff afterwards
 		long diff = 0;
@@ -496,7 +423,7 @@ public class Node {
 		if (parent == null || getEntryCount() > 2) {
 			//no merging required
 			//value exists --> remove it
-			return removeEntry(hcPos, pinToDelete, dims);
+			return removeEntry(hcPos, pinToDelete, dims, tree);
 		}
 
 		//okay, at his point we have a post that matches and (since it matches) we need to remove
@@ -524,7 +451,7 @@ public class Node {
 		}
 
 		long[] newPost = new long[dims];
-		RefsLong.arraycopy(key, 0, newPost, 0, key.length);
+		Bits.arraycopy(key, 0, newPost, 0, key.length);
 
 		long posInParent = PhTreeHelper.posInArray(key, parent.getPostLen());
 		int pinInParent = parent.getPosition(posInParent, dims);
@@ -537,50 +464,15 @@ public class Node {
 
 			//update parent, the position is the same
 			//we use newPost as Infix
-			parent.replaceEntryWithSub(pinInParent, posInParent, newPost, sub2);
+			parent.replaceEntryWithSub(pinInParent, posInParent, newPost, sub2, tree);
 		} else {
 			//this is also a post
 			getEntryByPIN(pin2, pos2, newPost);
-			parent.replaceSubWithPost(pinInParent, posInParent, newPost, val2);
+			parent.replaceSubWithPost(pinInParent, posInParent, newPost, val2, tree);
 		}
 
-		discardNode();
+		discardNode(tree);
 		return valueToDelete;
-	}
-
-	private void mergeIntoParentNt(long[] key, Node parent) {
-		int dims = key.length;
-
-		//check if merging is necessary (check children count || isRootNode)
-		if (parent == null || getEntryCount() > 2) {
-			//no merging required
-			//value exists --> remove it
-			return;
-		}
-
-		//okay, at his point we have a post that matches and (since it matches) we need to remove
-		//the local node because it contains at most one other entry and it is not the root node.
-		PhIterator64<Object> iter = ntIterator(dims);
-		NtEntry<Object> nte = iter.nextEntryReuse();
-
-		long posInParent = PhTreeHelper.posInArray(key, parent.getPostLen());
-		int pinInParent = parent.getPosition(posInParent, dims);
-		if (nte.getValue() instanceof Node) {
-			long[] newPost = nte.getKdKey();
-			//connect sub to parent
-			Node sub2 = (Node) nte.getValue();
-			int newInfixLen = getInfixLen() + 1 + sub2.getInfixLen();
-			sub2.setInfixLen(newInfixLen);
-
-			//update parent, the position is the same
-			//we use newPost as Infix
-			parent.replaceEntryWithSub(pinInParent, posInParent, newPost, sub2);
-		} else {
-			//this is also a post
-			parent.replaceSubWithPost(pinInParent, posInParent, nte.getKdKey(), nte.getValue());
-		}
-
-		discardNode();
 	}
 
 	/**
@@ -588,12 +480,7 @@ public class Node {
 	 * @param hcPos HC pos
 	 * @return The sub node or null.
 	 */
-	Object getEntryByPIN(int posInNode, long hcPos, long[] postBuf) {
-		if (isNT()) {
-			//For example for knnSearches!!!!!
-			return ntGetEntry(hcPos, postBuf, null);
-		}
-		
+	private Object getEntryByPIN(int posInNode, long hcPos, long[] postBuf) {
 		Object o = values[posInNode];
 		if (o instanceof Node) {
 			getInfixOfSub(posInNode, hcPos, postBuf);
@@ -676,11 +563,7 @@ public class Node {
 	 * @param newKey new key
 	 * @param value new value
 	 */
-	private void writeEntry(int pin, long hcPos, long[] newKey, Object value) {
-		if (isNT()) {
-			ntPut(hcPos, newKey, value);
-			return;
-		}
+	private void writeEntry(int pin, long hcPos, long[] newKey, Object value, PhTree13<?> tree) {
 		int dims = newKey.length;
 		int offsIndex = getBitPosIndex();
 		int offsKey;
@@ -715,25 +598,21 @@ public class Node {
 		return values[pin];
 	}
 
-	void replaceEntryWithSub(int posInNode, long hcPos, long[] infix, Node newSub) {
-		if (isNT()) {
-			ntReplaceEntry(hcPos, infix, newSub);
-			return;
-		}
-		//TODO during insert we wounldn't need to rewrite the infix, only the infix-flag 
+	private void replaceEntryWithSub(int posInNode, long hcPos, long[] infix, Node newSub, PhTree13<?> tree) {
+		//TODO during insert we wounldn't need to rewrite the infix, only the infix-flag
 		//     would need to be set...
 		writeSubInfix(posInNode, hcPos, infix, newSub.requiresInfix());
 		values[posInNode] = newSub;
 	}
 	
-	void writeSubInfix(int pin, long hcPos, long[] infix, boolean subRequiresInfix) {
+	private void writeSubInfix(int pin, long hcPos, long[] infix, boolean subRequiresInfix) {
 		//We should write the 1-bit hcPos here, even with infix-len==0.
 		replacePost(pin, hcPos, infix);
 		int dims = infix.length;
 		int subInfoOffs = pinToOffsBitsData(pin, hcPos, dims) + dims*postLenStored() - 1;
 		writeSubInfixInfo(ba, subInfoOffs, subRequiresInfix);
 	}
-	
+
 	private void writeSubInfixInfo(long[] ba, int subInfoOffs, boolean subRequiresInfix) {
 		//-> Should work for AHC and LHC with (offs+postLen-1)
 		
@@ -745,89 +624,23 @@ public class Node {
 	private boolean hasSubInfix(int subInfoOffs, int dims) {
 		return Bits.getBit(ba, subInfoOffs + dims*postLenStored() - 1);
 	}
-	
-	boolean hasSubInfixNI(long[] infix) {
-		//TODO reenable? But we also need to write it...
-		//return (infix[infix.length-1] & 1L) != 0;
-		return true;
-	}
-	
+
 	/**
 	 * Replace a sub-node with a postfix, for example if the current sub-node is removed, 
 	 * it may have to be replaced with a post-fix.
 	 */
-	void replaceSubWithPost(int pin, long hcPos, long[] key, Object value) {
-		if (isNT()) {
-			ntReplaceEntry(hcPos, key, value);
-			return;
-		}
+	private void replaceSubWithPost(int pin, long hcPos, long[] key, Object value, PhTree13<?> tree) {
 		values[pin] = value;
 		replacePost(pin, hcPos, key);
 	}
 
-	Object ntReplaceEntry(long hcPos, long[] kdKey, Object value) {
-		//We use 'null' as parameter to indicate that we want replacement, rather than splitting,
-		//if the value exists.
-		return NodeTreeV13.addEntry(ind, hcPos, kdKey, value, null);
-	}
-	
-	/**
-	 * General contract:
-	 * Returning a value or NULL means: Value was replaced, no change in counters
-	 * Returning a Node means: Traversal not finished, no change in counters
-	 * Returning null means: Insert successful, please update global entry counter
-	 * 
-	 * Node entry counters are updated internally by the operation
-	 * Node-counting is done by the NodePool.
-	 * 
-	 * @param hcPos HC pos
-	 * @param kdKey main key
-	 * @return see above
-	 */
-	Object ntPut(long hcPos, long[] kdKey, Object value) {
-		return NodeTreeV13.addEntry(ind, hcPos, kdKey, value, this);
-	}
-	
-	/**
-	 * General contract:
-	 * Returning a value or NULL means: Value was removed, please update global entry counter
-	 * Returning a Node means: Traversal not finished, no change in counters
-	 * Returning null means: Entry not found, no change in counters
-	 * 
-	 * Node entry counters are updated internally by the operation
-	 * Node-counting is done by the NodePool.
-	 * 
-	 * @param hcPos HC pos
-	 * @param dims dimensionality
-	 * @return see above
-	 */
-	Object ntRemoveAnything(long hcPos, int dims) {
-    	return NodeTreeV13.removeEntry(ind, hcPos, dims, null, null, null, null);
-	}
 
-	Object ntRemoveEntry(long hcPos, long[] key, long[] newKey, int[] insertRequired) {
-    	return NodeTreeV13.removeEntry(ind, hcPos, key.length, key, newKey, insertRequired, this);
-	}
-
-	Object ntGetEntry(long hcPos, long[] outKey, long[] valTemplate) {
-		return NodeTreeV13.getEntry(ind(), hcPos, outKey, null, null);
-	}
-
-	Object ntGetEntryIfMatches(long hcPos, long[] keyToMatch) {
-		return NodeTreeV13.getEntry(ind(), hcPos, null, keyToMatch, this);
-	}
-
-	int ntGetSize() {
-		return getEntryCount();
-	}
-	
-
-	private void switchLhcToAhcAndGrow(int oldEntryCount, int dims) {
+	private void switchLhcToAhcAndGrow(int oldEntryCount, int dims, PhTree13<?> tree) {
 		int posOfIndex = getBitPosIndex();
 		int posOfData = posToOffsBitsDataAHC(0, posOfIndex, dims);
 		setAHC( true );
-		long[] bia2 = Bits.arrayCreate(calcArraySizeTotalBits(oldEntryCount+1, dims));
-		Object [] v2 = Refs.arrayCreate(1<<dims);
+		long[] bia2 = tree.longPool().arrayCreate(calcArraySizeTotalBits(oldEntryCount+1, dims));
+		Object [] v2 = tree.objPool().arrayCreate(1<<dims);
 		//Copy only bits that are relevant. Otherwise we might mess up the not-null table!
 		Bits.copyBitsLeft(ba, 0, bia2, 0, posOfIndex);
 		int postLenTotal = dims*postLenStored(); 
@@ -839,16 +652,16 @@ public class Node {
 					postLenTotal);
 			v2[p2] = values[i];
 		}
-		ba = Bits.arrayReplace(ba, bia2);
-		values = Refs.arrayReplace(values, v2);
+		ba = tree.longPool().arrayReplace(ba, bia2);
+		values = tree.objPool().arrayReplace(values, v2);
 	}
 	
 	
-	private Object switchAhcToLhcAndShrink(int oldEntryCount, int dims, long hcPosToRemove) {
+	private Object switchAhcToLhcAndShrink(int oldEntryCount, int dims, long hcPosToRemove, PhTree13<?> tree) {
 		Object oldEntry = null;
 		setAHC( false );
-		long[] bia2 = Bits.arrayCreate(calcArraySizeTotalBits(oldEntryCount-1, dims));
-		Object[] v2 = Refs.arrayCreate(oldEntryCount-1);
+		long[] bia2 = tree.longPool().arrayCreate(calcArraySizeTotalBits(oldEntryCount-1, dims));
+		Object[] v2 = tree.objPool().arrayCreate(oldEntryCount-1);
 		int oldOffsIndex = getBitPosIndex();
 		int oldOffsData = oldOffsIndex + (1<<dims)*INN_HC_WIDTH;
 		//Copy only bits that are relevant. Otherwise we might mess up the not-null table!
@@ -872,8 +685,8 @@ public class Node {
 				n++;
 			}
 		}
-		ba = Bits.arrayReplace(ba, bia2);
-		values = Refs.arrayReplace(values, v2);
+		ba = tree.longPool().arrayReplace(ba, bia2);
+		values = tree.objPool().arrayReplace(values, v2);
 		return oldEntry;
 	}
 	
@@ -885,7 +698,7 @@ public class Node {
 	 * @param key new key
 	 * @param value new value
 	 */
-	void addPostPIN(long hcPos, int pin, long[] key, Object value) {
+	void addPostPIN(long hcPos, int pin, long[] key, Object value, PhTree13<?> tree) {
 		final int dims = key.length;
 		final int bufEntryCnt = getEntryCount();
 		//decide here whether to use hyper-cube or linear representation
@@ -894,17 +707,9 @@ public class Node {
 		//    For one dimension, both need one additional bit to indicate either
 		//    null/not-null (hypercube, actually two bit) or to indicate the index. 
 
-		if (!isNT() && shouldSwitchToNT(bufEntryCnt)) {
-			ntBuild(bufEntryCnt, dims, key);
-		}
-		if (isNT()) {
-			ntPut(hcPos, key, value);
-			return;
-		}
-
 		//switch representation (HC <-> Linear)?
 		if (!isAHC() && shouldSwitchToAHC(bufEntryCnt + 1, dims)) {
-			switchLhcToAhcAndGrow(bufEntryCnt, dims);
+			switchLhcToAhcAndGrow(bufEntryCnt, dims, tree);
 			//no need to update pin now, we are in HC now.
 		}
 
@@ -923,7 +728,7 @@ public class Node {
 			pin = -(pin+1);
 
 			//resize array
-			ba = Bits.arrayEnsureSize(ba, calcArraySizeTotalBits(bufEntryCnt+1, dims));
+			ba = tree.longPool().arrayEnsureSize(ba, calcArraySizeTotalBits(bufEntryCnt+1, dims));
 			long[] ia = ba;
 			int offs = pinToOffsBitsLHC(pin, offsIndex, dims);
 			Bits.insertBits(ia, offs, IK_WIDTH(dims) + dims*postLenStored);
@@ -935,180 +740,10 @@ public class Node {
 				Bits.writeArray(ia, offs, postLenStored(), key[i]);
 				offs += postLenStored();
 			}
-			values = Refs.insertSpaceAtPos(values, pin, bufEntryCnt+1);
+			values = tree.objPool().insertSpaceAtPos(values, pin, bufEntryCnt+1);
 			values[pin] = value;
 		}
 	}
-
-	void postToNI(int startBit, long[] outKey, long hcPos, long[] prefix, long mask) {
-		for (int d = 0; d < outKey.length; d++) {
-			outKey[d] = Bits.readArray(ba, startBit, postLenStored()) | (prefix[d] & mask);
-			startBit += postLenStored();
-		}
-	}
-
-	void postFromNI(long[] ia, int startBit, long key[]) {
-		//insert postifx
-		for (int d = 0; d < key.length; d++) {
-			Bits.writeArray(ia, startBit + postLenStored() * d, postLenStored(), key[d]);
-		}
-	}
-
-	void infixFromNI(long[] ba, int startBit, long[] key, boolean subRequiresInfix) {
-		//insert infix:
-		for (int i = 0; i < key.length; i++) {
-			Bits.writeArray(ba, startBit, postLenStored(), key[i]);
-			startBit += postLenStored();
-		}
-		int subInfoOffs = startBit-1; 
-		writeSubInfixInfo(ba, subInfoOffs, subRequiresInfix);
-	}
-
-	/**
-	 * WARNING: This is overloaded in subclasses of Node.
-	 * @return Index.
-	 */
-	NtNode<Object> createNiIndex(int dims) {
-		return NtNode.createRoot(dims);
-	}
-	
-	private void ntBuild(int bufEntryCnt, int dims, long[] prefix) {
-		//Migrate node to node-index representation
-		if (ind != null || isNT()) {
-			throw new IllegalStateException();
-		}
-		ind = createNiIndex(dims);
-
-		long prefixMask = mask1100(postLenStored());
-		
-		//read posts 
-		if (isAHC()) {
-			int oldOffsIndex = getBitPosIndex();
-			int oldPostBitsVal = posToOffsBitsDataAHC(0, oldOffsIndex, dims);
-			int postLenTotal = dims*postLenStored();
-			final long[] buffer = new long[dims];
-			for (int i = 0; i < (1L<<dims); i++) {
-				Object o = values[i];
-				if (o == null) {
-					continue;
-				} 
-				int dataOffs = oldPostBitsVal + i*postLenTotal;
-				postToNI(dataOffs, buffer, i, prefix, prefixMask);
-				//We use 'null' as parameter to indicate that we want 
-				//to skip checking for splitNode or increment of entryCount
-				NodeTreeV13.addEntry(ind, i, buffer, o, null);
-			}
-		} else {
-			int offsIndex = getBitPosIndex();
-			int dataOffs = pinToOffsBitsLHC(0, offsIndex, dims);
-			int postLenTotal = dims*postLenStored();
-			final long[] buffer = new long[dims];
-			for (int i = 0; i < bufEntryCnt; i++) {
-				long p2 = Bits.readArray(ba, dataOffs, IK_WIDTH(dims));
-				dataOffs += IK_WIDTH(dims);
-				Object e = values[i];
-				postToNI(dataOffs, buffer, p2, prefix, prefixMask);
-				//We use 'null' as parameter to indicate that we want 
-				//to skip checking for splitNode or increment of entryCount
-				NodeTreeV13.addEntry(ind, p2, buffer, e, null);
-				dataOffs += postLenTotal;
-			}
-		}
-
-		setAHC(false);
-		ba = Bits.arrayTrim(ba, calcArraySizeTotalBitsNt());
-		values = Refs.arrayReplace(values, null); 
-	}
-
-	/**
-	 * 
-	 * @param dims dimensions
-	 * @param posToRemove position to remove
-	 * @return Previous value if post was removed
-	 */
-	private Object ntDeconstruct(int dims, long posToRemove) {
-		//Migrate node to node-index representation
-		if (ind == null || !isNT()) {
-			throw new IllegalStateException();
-		}
-
-		int entryCountNew = ntGetSize() - 1;
-		decEntryCount();
-
-		//calc node mode.
-		boolean shouldBeAHC = useAHC(entryCountNew, dims);
-		setAHC(shouldBeAHC);
-
-
-		Object oldValue = null;
-		int offsIndex = getBitPosIndex();
-		long[] bia2 = Bits.arrayCreate(calcArraySizeTotalBits(entryCountNew, dims));
-		//Copy only bits that are relevant. Otherwise we might mess up the not-null table!
-		Bits.copyBitsLeft(ba, 0, bia2, 0, offsIndex);
-		int postLenTotal = dims*postLenStored();
-		if (shouldBeAHC) {
-			//HC mode
-			Object[] v2 = Refs.arrayCreate(1<<dims);
-			int startBitData = posToOffsBitsDataAHC(0, offsIndex, dims);
-			PhIterator64<Object> it = ntIterator(dims);
-			while (it.hasNext()) {
-				NtEntry<Object> e = it.nextEntryReuse();
-				long pos = e.key();
-				if (pos == posToRemove) {
-					//skip the item that should be deleted.
-					oldValue = e.value();
-					v2[(int) pos] = null;
-					continue;
-				}
-				int p2 = (int) pos;
-				int offsBitData = startBitData + postLenStored() * dims * p2;
-				if (e.value() instanceof Node) {
-					//subnode
-					Node node = (Node) e.value();
-					infixFromNI(bia2, offsBitData, e.getKdKey(), node.requiresInfix());
-				} else {
-					postFromNI(bia2, offsBitData, e.getKdKey());
-				}
-				v2[p2] = e.value();
-			}
-			ba = Bits.arrayReplace(ba, bia2);
-			values = Refs.arrayReplace(values, v2);
-		} else {
-			//LHC mode
-			Object[] v2 = Refs.arrayCreate(entryCountNew);
-			int n=0;
-			PhIterator64<Object> it = ntIterator(dims);
-			int entryPosLHC = offsIndex;
-			while (it.hasNext()) {
-				NtEntry<Object> e = it.nextEntryReuse();
-				long pos = e.key();
-				if (pos == posToRemove) {
-					//skip the item that should be deleted.
-					oldValue = e.value();
-					continue;
-				}
-				//write hc-key
-				Bits.writeArray(bia2, entryPosLHC, IK_WIDTH(dims), pos);
-				entryPosLHC += IK_WIDTH(dims);
-				v2[n] = e.value();
-				if (e.value() instanceof Node) {
-					Node node = (Node) e.getValue();
-					infixFromNI(bia2, entryPosLHC, e.getKdKey(), node.requiresInfix());
-				} else {
-					postFromNI(bia2, entryPosLHC, e.getKdKey());
-				}
-				entryPosLHC += postLenTotal;
-				n++;
-			}
-			ba = Bits.arrayReplace(ba, bia2);
-			values = Refs.arrayReplace(values, v2);
-		}			
-
-		NtNodePool.offer(ind);
-		ind = null;
-		return oldValue;
-	}
-
 
 	/**
 	 * Get post-fix.
@@ -1200,90 +835,10 @@ public class Node {
 		return true;
 	}
 
-	
-	boolean checkAndApplyInfixNt(int infixLen, long[] postFix, long[] valTemplate, 
-			long[] rangeMin, long[] rangeMax) {
-		//first check if node-prefix allows sub-node to contain any useful values
-
-		if (PhTreeHelper.DEBUG) {
-			N_GOOD++;
-			//Ensure that we never enter this method if the node cannot possibly contain a match.
-			long maskClean = mask1100(getPostLen());
-			for (int dim = 0; dim < valTemplate.length; dim++) {
-				if ((valTemplate[dim] & maskClean) > rangeMax[dim] || 
-						(valTemplate[dim] | ~maskClean) < rangeMin[dim]) {
-					if (getPostLen() < 63) {
-						System.out.println("N-CAAI: " + ++N + " / " + N_GOOD);
-						throw new IllegalStateException();
-					}
-					//ignore, this happens with negative values.
-					//return false;
-				}
-			}
-		}
-		
-		if (!hasSubInfixNI(postFix)) {
-			return true;
-		}
-
-		//assign infix
-		//Shift in two steps in case they add up to 64.
-		long maskClean = mask1100(postLenStored());
-
-		//first, clean trailing bits
-		//Mask for comparing the tempVal with the ranges, except for bit that have not been
-		//extracted yet.
-		long compMask = mask1100(postLenStored() - infixLen);
-		for (int dim = 0; dim < valTemplate.length; dim++) {
-			long inFull = (valTemplate[dim] & maskClean) | postFix[dim];
-			long in = inFull & compMask;
-			if (in > rangeMax[dim] || in < (rangeMin[dim]&compMask)) {
-				return false;
-			}
-			valTemplate[dim] = inFull;
-		}
-
-		return true;
-	}
-
-	public static long mask1100(int zeroBits) {
+	private static long mask1100(int zeroBits) {
 		return zeroBits == 64 ? 0 : ((-1L) << zeroBits);
 	}
 	
-	/**
-	 * Get post-fix.
-	 * @param hcPos HC pos
-	 * @param value current value of entry
-	 * @param result result
-	 * @param valTemplate The entry to check.
-	 * @param rangeMin After the method call, this contains the postfix if the postfix matches the
-	 * range. Otherwise it contains only part of the postfix.
-	 * @param rangeMax see rangeMin
-	 * @return NodeEntry if the postfix matches the range, otherwise null.
-	 */
-	@SuppressWarnings("unchecked")
-	<T>  boolean checkAndGetEntryNt(long hcPos, Object value, PhEntry<T> result, 
-			long[] valTemplate, long[] rangeMin, long[] rangeMax) {
-		if (value instanceof Node) {
-			Node sub = (Node) value;
-			if (!checkAndApplyInfixNt(sub.getInfixLen(), result.getKey(), valTemplate, 
-					rangeMin, rangeMax)) {
-				return false;
-			}
-			result.setNodeInternal(sub);
-		} else {
-			long[] inKey = result.getKey();
-			for (int i = 0; i < inKey.length; i++) {
-				long k = inKey[i];
-				if (k < rangeMin[i] || k > rangeMax[i]) {
-					return false;
-				}
-			}
-			result.setValueInternal((T) value);
-		}
-		return true;
-	}
-
 	private void applyHcPos(long hcPos, long[] valTemplate) {
 		PhTreeHelper.applyHcPos(hcPos, getPostLen(), valTemplate);
 	}
@@ -1306,21 +861,12 @@ public class Node {
 	}
 
 
-	Object removeEntry(long hcPos, int posInNode, final int dims) {
+	private Object removeEntry(long hcPos, int posInNode, final int dims, PhTree13<?> tree) {
 		final int bufEntryCnt = getEntryCount();
-		if (isNT()) {
-			if (shouldSwitchFromNtToHC(bufEntryCnt)) {
-				return ntDeconstruct(dims, hcPos);
-			}
-			Object o = ntRemoveAnything(hcPos, dims);
-			decEntryCount();
-			return o;
-		}
-		
 		//switch representation (HC <-> Linear)?
 		if (isAHC() && shouldSwitchToLHC(bufEntryCnt, dims)) {
 			//revert to linearized representation, if applicable
-			Object oldVal = switchAhcToLhcAndShrink(bufEntryCnt, dims, hcPos);
+			Object oldVal = switchAhcToLhcAndShrink(bufEntryCnt, dims, hcPos, tree);
 			decEntryCount();
 			return oldVal;
 		}			
@@ -1338,10 +884,10 @@ public class Node {
 			int posBit = pinToOffsBitsLHC(posInNode, offsIndex, dims);
 			Bits.removeBits(ba, posBit, IK_WIDTH(dims) + dims*postLenStored());
 			//shrink array
-			ba = Bits.arrayTrim(ba, calcArraySizeTotalBits(bufEntryCnt-1, dims));
+			ba = tree.longPool().arrayTrim(ba, calcArraySizeTotalBits(bufEntryCnt-1, dims));
 			//values:
 			oldVal = values[posInNode]; 
-			values = Refs.removeSpaceAtPos(values, posInNode, bufEntryCnt-1);
+			values = tree.objPool().removeSpaceAtPos(values, posInNode, bufEntryCnt-1);
 		}
 
 		decEntryCount();
@@ -1362,11 +908,6 @@ public class Node {
 	 */
 	void setAHC(boolean b) {
 		isHC = b;
-	}
-
-
-	boolean isNT() {
-		return ind != null;
 	}
 
 
@@ -1410,7 +951,7 @@ public class Node {
 		return offsIndex + (IK_WIDTH(dims) + postLenStored() * dims) * pin;
 	}
 	
-	int pinToOffsBitsData(int pin, long hcPos, int dims) {
+	private int pinToOffsBitsData(int pin, long hcPos, int dims) {
 		int offsIndex = getBitPosIndex();
 		if (isAHC()) {
 			return posToOffsBitsDataAHC(hcPos, offsIndex, dims);
@@ -1431,17 +972,10 @@ public class Node {
 			int posInt = (int) hcPos;  //Hypercube can not be larger than 2^31
 			return (values[posInt] != null) ? posInt : -(posInt)-1;
 		} else {
-			if (isNT()) {
-				//For NI, this value is not used, because checking for presence is quite 
-				//expensive. However, we have to return a positive value to avoid abortion
-				//of search (negative indicates that no value exists). It is hack though...
-				return Integer.MAX_VALUE;
-			} else {
-				//linearized cube
-				int offsInd = getBitPosIndex();
-				return Bits.binarySearch(ba, offsInd, getEntryCount(), hcPos, IK_WIDTH(dims), 
-						dims * postLenStored());
-			}
+			//linearized cube
+			int offsInd = getBitPosIndex();
+			return Bits.binarySearch(ba, offsInd, getEntryCount(), hcPos, IK_WIDTH(dims),
+					dims * postLenStored());
 		}
 	}
 
@@ -1453,7 +987,7 @@ public class Node {
 		return getInfixLen() > 0;
 	}
 
-	int infixLenStored() {
+	private int infixLenStored() {
 		return infixLenStored;
 	}
 
@@ -1467,18 +1001,6 @@ public class Node {
 
 	public int postLenStored() {
 		return postLenStored;
-	}
-	
-	NtNode<Object> ind() {
-		return ind;
-	}
-
-    PhIterator64<Object> ntIterator(int dims) {
-        return new NtIteratorMinMax<>(dims).reset(ind, Long.MIN_VALUE, Long.MAX_VALUE);
-    }
-
-    NtIteratorMask<Object> ntIteratorWithMask(int dims, long maskLower, long maskUpper) {
-		return new NtIteratorMask<>(dims).reset(ind, maskLower, maskUpper);
 	}
 
 	Object[] values() {
