+++ src/main/java/ch/ethz/globis/phtree/v13/NodeIteratorFullNoGC.java
@@ -1,17 +1,26 @@
 /*
  * Copyright 2011-2016 ETH Zurich. All Rights Reserved.
  * Copyright 2016-2018 Tilmann Zäschke. All Rights Reserved.
+ * Copyright 2019 Improbable. All rights reserved.
  *
- * This software is the proprietary information of ETH Zurich
- * and Tilmann Zäschke.
- * Use is subject to license terms.
+ * This file is part of the PH-Tree project.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package ch.ethz.globis.phtree.v13;
 
-import ch.ethz.globis.pht64kd.MaxKTreeI.NtEntry;
 import ch.ethz.globis.phtree.PhEntry;
 import ch.ethz.globis.phtree.PhFilter;
-import ch.ethz.globis.phtree.v13.nt.NtIteratorMinMax;
 
 
 
@@ -29,11 +38,9 @@ public class NodeIteratorFullNoGC<T> {
 	
 	private final int dims;
 	private boolean isHC;
-	private boolean isNI;
 	private long next = -1;
 	private Node node;
 	private int currentOffsetKey;
-	private NtIteratorMinMax<Object> ntIterator;
 	private int nMaxEntries;
 	private int nEntriesFound = 0;
 	private int postEntryLenLHC;
@@ -55,12 +62,7 @@ public class NodeIteratorFullNoGC<T> {
 	
 	/**
 	 * 
-	 * @param node
-	 * @param rangeMin The minimum value that any found value should have. If the found value is
-	 *  lower, the search continues.
-	 * @param rangeMax
-	 * @param lower The minimum HC-Pos that a value should have.
-	 * @param upper
+	 * @param node node
 	 * @param checker result verifier, can be null.
 	 */
 	private void reinit(Node node, PhFilter checker) {
@@ -70,20 +72,12 @@ public class NodeIteratorFullNoGC<T> {
 	
 		this.node = node;
 		this.isHC = node.isAHC();
-		this.isNI = node.isNT();
 		nMaxEntries = node.getEntryCount();
 		
 		
 		//Position of the current entry
-		if (isNI) {
-			if (ntIterator == null) {
-				ntIterator = new NtIteratorMinMax<>(dims);
-			}
-			ntIterator.reset(node.ind(), 0, Long.MAX_VALUE);
-		} else {
-			currentOffsetKey = node.getBitPosIndex();
-			postEntryLenLHC = Node.IK_WIDTH(dims)+dims*node.postLenStored();
-		}
+		currentOffsetKey = node.getBitPosIndex();
+		postEntryLenLHC = Node.IK_WIDTH(dims)+dims*node.postLenStored();
 	}
 
 	/**
@@ -143,11 +137,6 @@ public class NodeIteratorFullNoGC<T> {
 
 
 	private void getNext(PhEntry<T> result) {
-		if (isNI) {
-			niFindNext(result);
-			return;
-		}
-
 		if (isHC) {
 			getNextAHC(result);
 		} else {
@@ -180,17 +169,6 @@ public class NodeIteratorFullNoGC<T> {
 		} while (!readValue(nEntriesFound-1, currentPos, result));
 	}
 	
-	private void niFindNext(PhEntry<T> result) {
-		while (ntIterator.hasNext()) {
-			NtEntry<Object> e = ntIterator.nextEntryReuse();
-			if (readValue(e.key(), e.getKdKey(), e.value(), result)) {
-				next = e.key();
-				return;
-			}
-		}
-		next = FINISHED;
-	}
-
 	void init(Node node, PhFilter checker) {
 		reinit(node, checker);
 	}
