+++ src/main/java/ch/ethz/globis/phtree/v13SynchedPool/Bits.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2011-2016 ETH Zurich. All Rights Reserved.
+ * Copyright 2016-2018 Tilmann Zäschke. All Rights Reserved.
+ *
+ * This software is the proprietary information of ETH Zurich
+ * and Tilmann Zäschke.
+ * Use is subject to license terms.
+ */
+package ch.ethz.globis.phtree.v13SynchedPool;
+
+import ch.ethz.globis.phtree.util.BitsLong;
+
+
+/**
+ * Bit-stream manipulation functions.
+ * 
+ * @author ztilmann
+ *
+ */
+public class Bits extends BitsLong {
+
+
+}
