+++ src/main/java/ch/ethz/globis/phtree/v13SynchedPool/NodeIteratorFullToList.java
@@ -1,27 +1,17 @@
 /*
  * Copyright 2011-2016 ETH Zurich. All Rights Reserved.
  * Copyright 2016-2018 Tilmann Zäschke. All Rights Reserved.
- * Copyright 2019 Improbable. All rights reserved.
  *
- * This file is part of the PH-Tree project.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * This software is the proprietary information of ETH Zurich
+ * and Tilmann Zäschke.
+ * Use is subject to license terms.
  */
-package ch.ethz.globis.phtree.v13us;
+package ch.ethz.globis.phtree.v13SynchedPool;
 
+import ch.ethz.globis.pht64kd.MaxKTreeI.NtEntry;
 import ch.ethz.globis.phtree.PhEntryDist;
-import ch.ethz.globis.phtree.v13us.PhQueryKnnHS.KnnResultList;
-
+import ch.ethz.globis.phtree.v13SynchedPool.PhQueryKnnHS.KnnResultList;
+import ch.ethz.globis.phtree.v13SynchedPool.nt.NtIteratorMinMax;
 
 
 /**
@@ -36,39 +26,34 @@ public class NodeIteratorFullToList<T> {
 	
 	private final int dims;
 	private Node node;
+	private NtIteratorMinMax<Object> ntIterator;
 	private long[] prefix;
 	private final long maxPos;
 
 
 	/**
-	 * 
+	 *
 	 * @param dims dimensions
 	 */
 	public NodeIteratorFullToList(int dims) {
 		this.dims = dims;
 		this.maxPos = (1L << dims) -1;
 	}
-	
-	/**
-	 * 
-	 * @param node node
-	 * @param resultList result buffer
-	 * @param prefix prefix
-	 */
+
 	private void reinit(Node node, KnnResultList<T> resultList, long[] prefix) {
 		this.node = node;
 		this.prefix = prefix;
-		
+
 		getAll(resultList);
 	}
 
-	
+
 	@SuppressWarnings("unchecked")
 	private void readValue(int posInNode, long hcPos, KnnResultList<T> resultList) {
 		if (node.values()[posInNode] == null) {
 			return;
 		}
-		PhEntryDist<T> result = resultList.phGetTempEntry(); 
+		PhEntryDist<T> result = resultList.phGetTempEntry();
 		long[] key = result.getKey();
 		System.arraycopy(prefix, 0, key, 0, prefix.length);
 		//The key is the current (and future) prefix as well as key of key-value
@@ -76,7 +61,7 @@ public class NodeIteratorFullToList<T> {
 		if (v == null) {
 			throw new IllegalStateException();
 		}
-		
+
 		if (v instanceof Node) {
 			result.setNodeInternal(v);
 		} else {
@@ -86,10 +71,10 @@ public class NodeIteratorFullToList<T> {
 		resultList.phOffer(result);
 	}
 
-	
+
 	@SuppressWarnings("unchecked")
 	private void readValue(long[] kdKey, Object value, KnnResultList<T> resultList) {
-		PhEntryDist<T> result = resultList.phGetTempEntry(); 
+		PhEntryDist<T> result = resultList.phGetTempEntry();
 		if (value instanceof Node) {
 			Node sub = (Node) value;
 			System.arraycopy(kdKey, 0, result.getKey(), 0, kdKey.length);
@@ -104,21 +89,23 @@ public class NodeIteratorFullToList<T> {
 
 
 	private void getAll(KnnResultList<T> resultList) {
-		if (node.isAHC()) {
+		if (node.isNT()) {
+			niFindNext(resultList);
+		} else if (node.isAHC()) {
 			getAllAHC(resultList);
 		} else {
 			getNextLHC(resultList);
 		}
 	}
-	
+
 	private void getAllAHC(KnnResultList<T> resultList) {
 		//while loop until 1 is found.
-		long currentPos = 0; 
+		long currentPos = 0;
 		do {
 			readValue((int) currentPos, currentPos, resultList);
 		} while (++currentPos <= maxPos);
 	}
-	
+
 	private void getNextLHC(KnnResultList<T> resultList) {
 		int currentOffsetKey = node.getBitPosIndex();
 		int postEntryLenLHC = Node.IK_WIDTH(dims)+dims*node.postLenStored();
@@ -130,6 +117,17 @@ public class NodeIteratorFullToList<T> {
 			currentPos = Bits.readArray(node.ba(), currentOffsetKey, Node.IK_WIDTH(dims));
 			currentOffsetKey += postEntryLenLHC;
 			readValue(nEntriesFound-1, currentPos, resultList);
+		}
+	}
+	
+	private void niFindNext(KnnResultList<T> resultList) {
+		if (ntIterator == null) {
+			ntIterator = new NtIteratorMinMax<>(dims);
+		}
+		ntIterator.reset(node.ind(), 0, Long.MAX_VALUE);
+		while (ntIterator.hasNext()) {
+			NtEntry<Object> e = ntIterator.nextEntryReuse();
+			readValue(e.getKdKey(), e.value(), resultList);
 		}
 	}
