+++ src/main/java/ch/ethz/globis/phtree/v13SynchedPool/NodeIteratorNoGC.java
@@ -1,27 +1,17 @@
 /*
  * Copyright 2011-2016 ETH Zurich. All Rights Reserved.
  * Copyright 2016-2018 Tilmann Zäschke. All Rights Reserved.
- * Copyright 2019 Improbable. All rights reserved.
  *
- * This file is part of the PH-Tree project.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * This software is the proprietary information of ETH Zurich
+ * and Tilmann Zäschke.
+ * Use is subject to license terms.
  */
-package ch.ethz.globis.phtree.v13us;
+package ch.ethz.globis.phtree.v13SynchedPool;
 
+import ch.ethz.globis.pht64kd.MaxKTreeI.NtEntry;
 import ch.ethz.globis.phtree.PhEntry;
 import ch.ethz.globis.phtree.PhFilter;
-
+import ch.ethz.globis.phtree.v13SynchedPool.nt.NtIteratorMask;
 
 
 /**
@@ -38,9 +28,11 @@ public class NodeIteratorNoGC<T> {
 	
 	private final int dims;
 	private boolean isHC;
+	private boolean isNI;
 	private long next;
 	private Node node;
 	private int currentOffsetKey;
+	private NtIteratorMask<Object> niIterator;
 	private int nMaxEntry;
 	private int nFound = 0;
 	private int postEntryLenLHC;
@@ -54,7 +46,7 @@ public class NodeIteratorNoGC<T> {
 	private PhFilter checker;
 
 	/**
-	 * 
+	 *
 	 * @param dims dimensions
 	 * @param valTemplate A null indicates that no values are to be extracted.
 	 */
@@ -62,9 +54,9 @@ public class NodeIteratorNoGC<T> {
 		this.dims = dims;
 		this.valTemplate = valTemplate;
 	}
-	
+
 	/**
-	 * 
+	 *
 	 * @param node node
 	 * @param rangeMin The minimum value that any found value should have. If the found value is
 	 *  lower, the search continues.
@@ -78,9 +70,10 @@ public class NodeIteratorNoGC<T> {
 		currentOffsetKey = 0;
 		nFound = 0;
 		this.checker = checker;
-	
+
 		this.node = node;
 		this.isHC = node.isAHC();
+		this.isNI = node.isNT();
 		nMaxEntry = node.getEntryCount();
 		if (!isHC) {
 			//Position of the current entry
@@ -95,6 +88,13 @@ public class NodeIteratorNoGC<T> {
 		if (dims > 6) {
 			initHCI();
 		}
+
+		if (isNI && !useNiHcIncrementer) {
+			if (niIterator == null) {
+				niIterator = new NtIteratorMask<>(dims);
+			}
+			niIterator.reset(node.ind(), maskLower, maskUpper);
+		}
 	}
 
 	private void initHCI() {
@@ -103,20 +103,31 @@ public class NodeIteratorNoGC<T> {
 		int nSetFilterBits = Long.bitCount(maskLower | ((~maskUpper) & maxHcAddr));
 		//nPossibleMatch = (2^k-x)
 		long nPossibleMatch = 1L << (dims - nSetFilterBits);
-		if (PhTree13us.HCI_ENABLED){
+		if (isNI) {
+			int nChild = node.ntGetSize();
+			int logNChild = Long.SIZE - Long.numberOfLeadingZeros(nChild);
+			//the following will overflow for k=60
+			boolean useHcIncrementer = (nChild > nPossibleMatch*(double)logNChild*2);
+			//DIM < 60 as safeguard against overflow of (nPossibleMatch*logNChild)
+			if (useHcIncrementer && PhTree13SP.HCI_ENABLED && dims < 50) {
+				useNiHcIncrementer = true;
+			} else {
+				useNiHcIncrementer = false;
+			}
+		} else if (PhTree13SP.HCI_ENABLED){
 			if (isHC) {
 				//nPossibleMatch < 2^k?
 				//PAPER
 				useHcIncrementer = nPossibleMatch*2 <= maxHcAddr;
 			} else {
 				int logNPost = Long.SIZE - Long.numberOfLeadingZeros(nMaxEntry) + 1+1;
-				useHcIncrementer = (nMaxEntry >= 2*nPossibleMatch*(double)logNPost); 
+				useHcIncrementer = (nMaxEntry >= 2*nPossibleMatch*(double)logNPost);
 			}
 		}
 	}
-	
+
 	/**
-	 * Advances the cursor. 
+	 * Advances the cursor.
 	 * @return TRUE iff a matching element was found.
 	 */
 	boolean increment(PhEntry<T> result) {
@@ -124,21 +135,21 @@ public class NodeIteratorNoGC<T> {
 	}
 
 	/**
-	 * 
+	 *
 	 * @return False if the value does not match the range, otherwise true.
 	 */
 	@SuppressWarnings("unchecked")
 	private boolean readValue(int pin, long pos, PhEntry<T> result) {
-		Object o = node.checkAndGetEntryPIN(pin, pos, valTemplate, result.getKey(), 
+		Object o = node.checkAndGetEntryPIN(pin, pos, valTemplate, result.getKey(),
 				rangeMin, rangeMax);
 		if (o == null) {
 			return false;
 		}
-		
+
 		if (o instanceof Node) {
 			Node sub = (Node) o;
 			//skip this for postLen>=63
-			if (checker != null && sub.getPostLen() < (PhTree13us.DEPTH_64-1) &&
+			if (checker != null && sub.getPostLen() < (PhTree13SP.DEPTH_64-1) &&
 					!checker.isValid(sub.getPostLen()+1, valTemplate)) {
 				return false;
 			}
@@ -153,6 +164,25 @@ public class NodeIteratorNoGC<T> {
 		return true;
 	}
 
+	private boolean readValue(long pos, Object value, PhEntry<T> result) {
+		if (!node.checkAndGetEntryNt(pos, value, result, valTemplate, rangeMin, rangeMax)) {
+			return false;
+		}
+
+		//subnode ?
+		if (value instanceof Node) {
+			Node sub = (Node) value;
+			//skip this for postLen>=63
+			if (checker != null && sub.getPostLen() < (PhTree13SP.DEPTH_64-1) &&
+					!checker.isValid(sub.getPostLen()+1, valTemplate)) {
+				return false;
+			}
+			return true;
+		}
+
+		return checker == null || checker.isValid(result.getKey());
+	}
+
 	private boolean getNextHCI(PhEntry<T> result) {
 		//Ideally we would switch between b-serch-HCI and incr-search depending on the expected
 		//distance to the next value.
@@ -162,7 +192,7 @@ public class NodeIteratorNoGC<T> {
 				//starting position
 				currentPos = maskLower;
 			} else {
-				currentPos = PhTree13us.inc(currentPos, maskLower, maskUpper);
+				currentPos = PhTree13SP.inc(currentPos, maskLower, maskUpper);
 				if (currentPos <= maskLower) {
 					return false;
 				}
@@ -177,6 +207,10 @@ public class NodeIteratorNoGC<T> {
 	}
 
 	private boolean getNext(PhEntry<T> result) {
+		if (isNI) {
+			return niFindNext(result);
+		}
+
 		if (useHcIncrementer) {
 			return getNextHCI(result);
 		} else if (isHC) {
@@ -185,10 +219,10 @@ public class NodeIteratorNoGC<T> {
 			return getNextLHC(result);
 		}
 	}
-	
+
 	private boolean getNextAHC(PhEntry<T> result) {
 		//while loop until 1 is found.
-		long currentPos = next == START ? maskLower-1 : next; 
+		long currentPos = next == START ? maskLower-1 : next;
 		while (true) {
 			currentPos++;  //pos w/o bit-offset
 			if (currentPos > maskUpper) {
@@ -201,7 +235,7 @@ public class NodeIteratorNoGC<T> {
 			}
 		}
 	}
-	
+
 	private boolean getNextLHC(PhEntry<T> result) {
 		while (true) {
 			if (++nFound > nMaxEntry) {
@@ -221,7 +255,58 @@ public class NodeIteratorNoGC<T> {
 			}
 		}
 	}
-	
+
+
+	private boolean niFindNext(PhEntry<T> result) {
+		return useNiHcIncrementer ? niFindNextHCI(result) : niFindNextIter(result);
+	}
+
+	private boolean niFindNextIter(PhEntry<T> result) {
+		while (niIterator.hasNext()) {
+			NtEntry<Object> e = niIterator.nextEntryReuse();
+			System.arraycopy(e.getKdKey(), 0, result.getKey(), 0, dims);
+			if (readValue(e.key(), e.value(), result)) {
+				next = e.getKey(); //This is required for kNN-adjusting of iterators
+				return true;
+			}
+		}
+		return false;
+	}
+
+	private boolean niFindNextHCI(PhEntry<T> result) {
+		//HCI
+		//repeat until we found a value inside the given range
+		long currentPos = next;
+		do {
+			if (currentPos != START && currentPos >= maskUpper) {
+				break;
+			}
+
+			if (currentPos == START) {
+				//starting position
+				currentPos = maskLower;
+			} else {
+				currentPos = PhTree13SP.inc(currentPos, maskLower, maskUpper);
+				if (currentPos <= maskLower) {
+					break;
+				}
+			}
+
+			Object v = node.ntGetEntry(currentPos, result.getKey(), valTemplate);
+			if (v == null) {
+				continue;
+			}
+
+			next = currentPos;
+
+			//read and check post-fix
+			if (readValue(currentPos, v, result)) {
+				return true;
+			}
+		} while (true);
+		return false;
+	}
+
 
 	private boolean checkHcPos(long pos) {
 		return ((pos | maskLower) & maskUpper) == pos;
@@ -231,11 +316,6 @@ public class NodeIteratorNoGC<T> {
 		return node;
 	}
 
-	/**
-	 * 
-	 * @param rangeMin min
-	 * @param rangeMax max
-	 */
 	private void calcLimits(long[] rangeMin, long[] rangeMax) {
 		//create limits for the local node. there is a lower and an upper limit. Each limit
 		//consists of a series of DIM bit, one for each dimension.
@@ -311,9 +391,15 @@ public class NodeIteratorNoGC<T> {
 			if (isHC) {
 				currentOffsetKey = node.getBitPosIndex();
 				next = START;
+			} else if (isNI) {
+				if (!useNiHcIncrementer) {
+					niIterator.adjustMinMax(maskLower, maskUpper);
+				} else {
+					next = START;
+				}
 			} else {
 				//LHC
-				if (this.next + PhTree13us.LHC_BINARY_SEARCH_THRESHOLD < this.maskLower) {
+				if (this.next + PhTree13SP.LHC_BINARY_SEARCH_THRESHOLD < this.maskLower) {
 					int pin = node.getPosition(maskLower, dims);
 					//If we don't find it we use the next following entry, i.e. -(pin+1)
 					pin = pin >= 0 ? pin : -(pin+1);
